requirements.txt
litellm>=1.40.0
langfuse>=2.0.0
chromadb>=0.5.0
openai>=1.30.0
tree-sitter>=0.22.0
tree-sitter-typescript>=0.23.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
gitpython>=3.1.43
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
google-cloud-aiplatform>=1.60.0
